保存交易
sendraw命令

+++ allpet.node/Node_Network_RPC.cs
@@ -25,14 +25,14 @@ namespace AllPet.Module
     }
     public partial class Module_Node : Module_MsgPack
     {
-        private ulong lastIndex = -1;
+        private long lastIndex = -1;
         private ulong GetLastIndex()
         {
             lock(this)
             {
                 lastIndex++;
             }
-            return lastIndex;
+            return (ulong)lastIndex;
         }
         public RPC_Result RPC_ListPeer(IList<MessagePackObject> _params)
         {
